+++ Beam/resources/views/dashboard/public.blade.php
@@ -24,7 +24,7 @@
             },
             created: function() {
                 this.$store.commit('changeSettings', {
-                    newGraph: true,
+                    newGraph: {{ json_encode(config('beam.pageview_graph_data_source') === 'snapshots') }}
                 })
             },
             provide: function() {